requirements.txt
pypdf
langchain
langchain-community
chromadb # Vector storage
pytest
boto3
sentence-transformers
langchain-huggingface
huggingface_hub
python-dotenv
pygame
PyQt6
pyaudio
SpeechRecognition
gtts
sounddevice
soundfile
numpy
openai-whisper
dataclasses
scholarly>=1.7.11
beautifulsoup4
requests
bibtexparser>=2.0.0b4
